계차수열

+++ SWEA/D3/Solution_D3_8016_홀수피라미드.java
@@ -11,11 +11,11 @@ public class Solution_D3_8016_홀수피라미드 {
 		for(int tc=1;tc<=T;tc++) {
 			st = new StringTokenizer(br.readLine(), " ");
 			
-			//long N = Long.parseLong(st.nextToken()); 
-			int N = Integer.parseInt(st.nextToken());
-			// 왼쪽 수열 점화식
-			int leftSequence = 1 + 2 * (N - 1) * (N - 1);
-			int rightSequence = 1 + 2 * (N - 1) * (N + 1);
+			long N = Long.parseLong(st.nextToken()); 
+			
+			// 왼쪽, 오른쪽 수열 점화식
+			long leftSequence = 1 + 2 * (N - 1) * (N - 1);
+			long rightSequence = 1 + 2 * (N - 1) * (N + 1);
 			
 			System.out.println("#" + tc + " " + leftSequence + " " + rightSequence);
 		}
